move it all into wallets
