+++ shared/stories/stories.js
@@ -13,7 +13,7 @@ import signup from '../login/signup/index.stories'
 import styles from '../styles/index.stories'
 import teams from '../teams/index.stories'
 import unlockFolders from '../unlock-folders/index.stories'
-import stellar from '../stellar/index.stories'
+import wallets from '../wallets/index.stories'
 
 const stories = {
   chat,
@@ -27,10 +27,10 @@ const stories = {
   search,
   settings,
   signup,
-  stellar,
   styles,
   teams,
   unlockFolders,
+  wallets,
 }
 
 export default stories
