+++ shared/wallets/wallet-list/index.js
